eval: extract the judge contract from its model call
The prompt, schema, and consistency guards move to a module with no product
imports, leaving only the model call behind, so the rubric can be reused
without pulling in provider plumbing.

The prompt text is unchanged, so results measured against it still apply.
